Fix a little warning.

+++ LunaDll/FileManager/SMBXFileManager.cpp
@@ -210,7 +210,7 @@ void SMBXWorldFileBase::PopulateEpisodeList()
             ep->episodeName = wldData.EpisodeTitle;
             ep->episodePath = epPath + "\\";
             ep->episodeWorldFile = wldName;
-            for (int i = 0; i < 5; i++)
+            for (size_t i = 0; i < 5; i++)
             {
                 if (i < wldData.nocharacter.size())
                 {
